\name{trifield-package}
\alias{trifield-package}
\alias{trifield}
\docType{package}
\title{
Assists in creating and plotting ternary fields
}
\description{
The package can project between unity-summed ternary vectors and binary vectors. An "apply" function is provided to apply a function over a ternary field, as is a function for constructing ternary grids. Some rudimentary plotting functions are also provided. This is quick and dirty research code I needed in the course of writing a manuscript. Take it as such.
}
\details{
\tabular{ll}{
Package: \tab trifield\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-24\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
Some simple routines, mainly aimed at making ternary fields and contour plots.
}
\author{
Tim Keitt <tkeitt@gmail.com>

Maintainer: Tim Keitt <tkeitt@gmail.com>
}
\keyword{ package }
